Update version of @bazel/worker to fix error reporting with workers enabled. (#95)
Currently the sass rules do not run properly with
`--spawn_strategy=worker` because the Sass build is
asynchronous and the worker utilities do not support
promises in previous versions. This causes the worker
to complete the action before the Sass compilation finishes.

Meaning that all failures are not properly reported back to
Bazel through the `worker_protocol`. Causing the sass rule
to silently fail with a message saying that the output has
not been generated.