refactor: do not write output files if compilation failed (#106)
By default, the CLI of Sass writes the output file even if compilation
failures have been reported. We don't want this behavior in the Bazel
action, as wirting the actual output file could let the compilation
action appear successful. Instead, if we donot write any file on error,
Bazel will never report the action as successful if an error occurred.